requirements.txt
asgiref==3.5.2
Django==4.1.3
django-geojson==4.0.0
django-leaflet==0.28.3
gunicorn==20.1.0
psycopg2==2.9.5
sqlparse==0.4.3
tzdata==2022.7
